Move lib to src folder :file_cabinet:
