+++ kinesis.js
@@ -1,6 +1,6 @@
-const {request} = require('./lib/kinesis-request');
-const {listStreams} = require('./lib/kinesis-api');
-const KinesisStream = require('./lib/kinesis-stream');
+const {request} = require('./src/kinesis-request');
+const {listStreams} = require('./src/kinesis-api');
+const KinesisStream = require('./src/kinesis-stream');
 
 module.exports.stream = options => new KinesisStream(options);
 module.exports.KinesisStream = KinesisStream;
