+++ src/kinesis-api.js
