+++ src/kinesis-request.js
