+++ src/kinesis-stream.js
